feat: multiple improvements
* Improve performance of view-all-disruptions page by retrieving operators and services on opening filters
* Improve site performance by using warming function to keep site server lambdas warm
* Use memoisation to prevent flickering of map markers on services consequence page
* Close account menu on selecting item
* Don't allow staff members to delete disruptions
* Increase refresh token age to 30 days

+++ site/pages/api/delete-disruption.api.ts
@@ -1,8 +1,9 @@
+import { PublishStatus } from "@create-disruptions-data/shared-ts/enums";
 import { NextApiRequest, NextApiResponse } from "next";
 import { ERROR_PATH } from "../../constants";
 import { deletePublishedDisruption, getDisruptionById } from "../../data/dynamo";
 import { redirectTo, redirectToError } from "../../utils/apiUtils";
-import { getSession } from "../../utils/apiUtils/auth";
+import { canPublish, getSession } from "../../utils/apiUtils/auth";
 import logger from "../../utils/logger";
 
 const deleteDisruption = async (req: NextApiRequest, res: NextApiResponse): Promise<void> => {
@@ -24,6 +25,10 @@ const deleteDisruption = async (req: NextApiRequest, res: NextApiResponse): Prom
             logger.error(`Disruption ${id} not found to delete`);
             redirectTo(res, ERROR_PATH);
             return;
+        }
+
+        if (!canPublish(session) && disruption.publishStatus !== PublishStatus.draft) {
+            throw new Error(`Insufficient permissions to delete disruption (${id})`);
         }
 
         await deletePublishedDisruption(disruption, id, session.orgId);
